congnghehiendaiapi/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==4.2.13
django-cloudinary-storage==0.3.0
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.0
PyMySQL==1.1.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
